faq-telegram-bot/requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0.0
langchain-gigachat>=0.3.12
